server/balancer_worker: check in progress operator
When an operator is in progress, we should not replace it.

For example, we have a region with 3 peers (A,B,C).  The balancer add
an operator to add a peer D and replace peer A.  Then, the region adds
the peer D and reports to PD that it has 4 peers (A,B,C,D). Now the
replica balancer will take over to remove a peer, and it may remove
peer D, which is not reasonable.

Reference Issues: [TiKV #1084](https://github.com/pingcap/tikv/issues/1084)