Add error codes (#1137)
* Add error codes so that clients can reliably understand errors. One error was added that allows a client to understand that a store is already tombstoned when remove is requested.

For extensibility error codes are text (reserving numbers is a pain).
The error code will show up in both a header and in the JSON body which can contain additional structured data about the error. This is in addition to sending back a detailed error message as is done now. The main goal was just to send back the error code in the header. Our ideas about the JSON body might change.