pdctl: support to specify multiple pd addrs seperated by commas
Signed-off-by: Shafreeck Sea <shafreeck@gmail.com>

Closes #1611

Example:
 $ pd-ctl -u 'http://192.168.10.1:2379,http://192.168.10.2:2379'

pd-ctl prefers to stick on a health endpoint until getting some error,
then it chooses a next endpoint in a round-robin manner and keeps using
it until next error.

It should be noticed that if current endpoint can not fail fast, it
maybe slow to get the result while the next endpoint is healthy, because
we were waiting for the current endpoint to return until it was timeout.