build: add tests for building docker image
Issue #2138 #2136 reference difficulties faced when our `docker build`
is broken. The commit attempts to mitigate the difficulties by adding a
sanity test in `make check` command. `make check` command is run in the
github workflow as well, so we would be able to locate any issues caused
due to a broken `docker build` quickly.

Signed-off-by: AnubhavUjjawal <anubhavujjawal@gmail.com>