server: allow the follower to handle QueryRegion (#9196)
ref tikv/pd#8690

Backport the server-side portion of #9196: branch the QueryRegion stream
handler on leadership. When served by a follower, read regions from the
follower's region syncer (s.cluster) and return regionNotFound if the
syncer is not running, mirroring the existing follower-handle behavior of
GetRegion / GetPrevRegion / GetRegionByID.

The client-side follower routing from #9196 is omitted since the router
client does not exist on release-8.5 and will be backported separately.

Signed-off-by: JmPotato <github@ipotato.me>