gc: Optimize GetAllKeyspacesGCStates for scenarios where there are too many keyspaces (#9777)
ref tikv/pd#8978

This PR supports loading keyspaces by smaller batches in `GetAllKeyspacesGCStates`. This avoids it througing errors due to etcd txn size limit when there are too many keyspaces.
Also makes `GetAllKeyspacesGCStates` support cancelling.
This is not the ultimate solution for tracking GC states changes. The best way is to implement a streaming API that watches the changes of the GC states. However, this PR provides a simpler fix to make it able to run.

Signed-off-by: MyonKeminta <MyonKeminta@users.noreply.github.com>